Spawn story in container based on its coordinates
Added a getter to the FileLoader class which makes the stories globally
accessible.

The mod uses the OnFillContainer event to hook into the process of
filling a container in the world. It then compares the coordinates of
the container with the list of stories. If a story has the same
coordinates as the container it will be spawned.

Spawning means, we create a (placeholder) item and drop it into the
container. The story's title and body are stored in the modData of the
newly created item.